Move more numbers to the params. Some preparations for the load_simulation_inputs function.
