+++ src/config.py
@@ -1,6 +1,8 @@
 import warnings
 from pathlib import Path
 
+import pandas as pd
+
 FAST_FLAG = "debug"
 """One of 'debug', 'verify', 'full'.
 
@@ -36,11 +38,5 @@ POPULATION_GERMANY = 83_000_000
 
 N_HOUSEHOLDS = 1_000_000
 
-SHARE_REFUSE_VACCINATION = 0.15
-"""Share of individuals refusing to be vaccinated.
-
-    80% of Germans are somewhat or definitely willing to be vaccinated.
-    12% are undecided. 8% are opposed to being vaccinated.
-    We assume that 15% will refuse to be vaccinated.
-    source: https://bit.ly/3c9mTgX (publication date: 2021-03-02)
-"""
+VERY_EARLY = pd.Timestamp("2020-01-01")
+VERY_LATE = pd.Timestamp("2022-12-31")
