+++ src/testing/task_get_and_plot_share_of_tests_for_symptomatics.py
@@ -27,6 +27,8 @@ OUT_PATH = BLD / "data" / "testing"
 @pytask.mark.produces(
     {
         "data": OUT_PATH / "characteristics_of_the_tested.csv",
+        "share_of_tests_for_symptomatics_series": OUT_PATH
+        / "share_of_tests_for_symptomatics_series.pkl",
         "mean_age": OUT_PATH / "mean_age_of_tested.png",
         "share_with_symptom_status": OUT_PATH
         / "share_of_tested_with_symptom_status.png",
@@ -69,6 +71,16 @@ def task_prepare_characteristics_of_the_tested(depends_on, produces):
     fig.savefig(produces["symptom_shares"])
     df = df.reset_index().rename(columns={"index": "date"})
     df.to_csv(produces["data"])
+
+    share_of_tests_for_symptomatics_series = df[
+        [
+            "share_symptomatic_lower_bound_extrapolated",
+            "share_symptomatic_among_known_extrapolated",
+        ]
+    ].mean(axis=1)
+    share_of_tests_for_symptomatics_series.to_pickle(
+        produces["share_of_tests_for_symptomatics_series"]
+    )
 
 
 def _clean_data(df):
